% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bnp_bernoulli.R
\name{anova_bnp_bernoulli}
\alias{anova_bnp_bernoulli}
\title{Fit a Bayesian nonparametric factorial ANOVA model with Bernoulli kernel.}
\usage{
anova_bnp_bernoulli(
  y,
  X,
  iter = 4000L,
  warmup = 2000L,
  seed = 1L,
  rho = 1,
  a = 1,
  b = 1,
  a2 = 2,
  b2 = 4,
  lb = min(y),
  ub = max(y)
)
}
\arguments{
\item{y}{An boolean response vector.}

\item{X}{A design matrix (full of integer covariates).}

\item{iter}{the total number of mcmc iterations (default 4000).}

\item{warmup}{the number of warmup mcmc iterations (default 2000).}

\item{seed}{the seed for random number generation (default 1L).}

\item{rho}{the hyperparameter \eqn{\rho} (default 1.0).}

\item{a}{the hyperparameter \eqn{a} (default 1.0).}

\item{b}{the hyperparameter \eqn{b} (default 1.0).}

\item{a2}{the hyperparameter \eqn{a_2} (default 2.0).}

\item{b2}{the hyperparameter \eqn{b_2} (default 4.0).}
}
\value{
An object of class 'anova_bnp_model'.
}
\description{
Fit a Bayesian nonparametric factorial ANOVA model with Bernoulli kernel.
}
