% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bnp_berpoi.R
\name{anova_bnp_berpoi}
\alias{anova_bnp_berpoi}
\title{Fit a Bayesian nonparametric factorial ANOVA model with BerPoi kernel.}
\usage{
anova_bnp_berpoi(
  y,
  X,
  iter = 4000L,
  warmup = 2000L,
  seed = 1L,
  rho = 1,
  a = 1,
  b = 1,
  a1 = 2,
  b1 = 4,
  alpha0 = 1,
  beta0 = 1,
  lb = min(y),
  ub = max(y)
)
}
\arguments{
\item{y}{An integer response vector.}

\item{X}{A design matrix (full of integer covariates).}

\item{iter}{the total number of mcmc iterations (default 4000).}

\item{warmup}{the number of warmup mcmc iterations (default 2000).}

\item{seed}{the seed for random number generation (default 1L).}

\item{rho}{the hyperparameter \eqn{\rho} (default 1.0).}

\item{a}{the hyperparameter \eqn{a} (default 1.0).}

\item{b}{the hyperparameter \eqn{b} (default 1.0).}

\item{a1}{the hyperparameter \eqn{a_1} (default 2.0).}

\item{b1}{the hyperparameter \eqn{b_1} (default 4.0).}

\item{alpha0}{the hyperparameter \eqn{\alpha_0} (default 1.0).}

\item{beta0}{the hyperparameter \eqn{\beta_0} (default 1.0).}
}
\value{
An object of class 'anova_bnp_model'.
}
\description{
Fit a Bayesian nonparametric factorial ANOVA model with BerPoi kernel.
}
