% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_01}
\alias{example_01}
\title{The 1st dataset used in the "Getting Started" vignette}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
\item{y}{a response variable that taking real values.}
\item{x1}{a simulated factor with levels 1 and 2.}
\item{x2}{a simulated factor with levels 1 and 2.}
}
}
\source{
Own elaboration
}
\usage{
example_01
}
\description{
The 1st dataset used in the "Getting Started" vignette
}
\keyword{datasets}
