% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_chain.R
\name{gamma_chain}
\alias{gamma_chain}
\title{Return the markov chain associated to the hypothesis vector}
\usage{
gamma_chain(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
A tibble with the markov chian associated to the hypothesis
vector, and the following variables:
\itemize{
\item iter: MCMC iteration.
\item variable: The name of the variable.
\item value: The value of the variable at iteration each iteration.
}
For more information about the meaning of each group, use \code{group_codes}.
}
\description{
Return the markov chain associated to the hypothesis vector
}
