% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_post.R
\name{gamma_post}
\alias{gamma_post}
\title{Return the posterior probability of each hypothesis vector}
\usage{
gamma_post(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
A tibble with the posterior probability of each hypothesis
(one per row), and the following variables:
\itemize{
\item prob: the posterior probability.
\item group\eqn{i}: TRUE if, under that hypothesis, the group 1 shares the
distribution of the reference group, for each \eqn{i = 2, 3, \ldots}
}
For more information about the meaning of each group, use \code{group_codes}.
}
\description{
Return the posterior probability of each hypothesis vector
}
