% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_codes.R
\name{group_codes}
\alias{group_codes}
\title{Return the factor levels associated with each group}
\usage{
group_codes(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
A tibble with the factor levels associated with each group
and the following variables:
\itemize{
\item group: the group id.
\item x\eqn{i}: the level of the \eqn{i}th factor for that  group,
for each \eqn{i = 1, 2, \ldots}
}
}
\description{
Return the factor levels associated with each group
}
