% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_post_interaction.R
\name{hypothesis_post_interaction}
\alias{hypothesis_post_interaction}
\title{Return the YYY statistic associated with each factor}
\usage{
hypothesis_post_interaction(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
A tibble with the following variables:
\itemize{
\item var1: the 1st factor id.
\item var2: the 2nd factor id.
\item prob: the YYY summary statistic.
}
}
\description{
Return the YYY statistic associated with each factor
}
