% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_post_simple.R
\name{hypothesis_post_simple}
\alias{hypothesis_post_simple}
\title{Return the XXX statistic associated with each factor}
\usage{
hypothesis_post_simple(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
A tibble with the following variables:
\itemize{
\item var1: factor id.
\item prob: the XXX summary statistic.
}
}
\description{
Return the XXX statistic associated with each factor
}
