% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_significance_probability.R
\name{joint_significance_probability}
\alias{joint_significance_probability}
\title{Compute the posterior probability of having at least one relevant factor}
\usage{
joint_significance_probability(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
The corresponding probability.
}
\description{
Compute the posterior probability of having at least one relevant factor
}
