% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_plot_interaction.R
\name{predictive_plot_interaction}
\alias{predictive_plot_interaction}
\title{Plot the posterior predictive pdf (interactions)}
\usage{
predictive_plot_interaction(fit, d1, d2)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}

\item{d1}{The 1st factor id.}

\item{d2}{The 2nd factor id.}
}
\value{
A ggplot2 plot.
}
\description{
Plot the posterior predictive pdf for each combination of 2 factors,
fixing the others at the reference level.
}
