% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_plot.R
\name{shift_plot}
\alias{shift_plot}
\title{Plot a shift function}
\usage{
shift_plot(fit, group1, group2)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}

\item{group}{The group id.}
}
\value{
A ggplot2 plot.
}
\description{
Plot the shift function for one particular group.
}
