% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_post.R
\name{shift_post}
\alias{shift_post}
\title{Return the shift functions}
\usage{
shift_post(fit)
}
\arguments{
\item{fit}{An object of class anova_bnp_model.}
}
\value{
A tibble with the following variables:
\itemize{
\item group the group id.
\item y the value of the response variable.
\item the shift function evaluated at y, given the group.
}
}
\description{
Return the shift functions
}
