% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bnp_bernoulli.R
\name{anova_bnp_bernoulli}
\alias{anova_bnp_bernoulli}
\title{Fit a Bayesian nonparametric factorial ANOVA model with Bernoulli kernel}
\usage{
anova_bnp_bernoulli(
  y,
  X,
  iter = 4000L,
  warmup = 2000L,
  seed = 1L,
  rho = 1,
  a = 1,
  b = 1,
  a2 = 2,
  b2 = 4,
  lb = min(y),
  ub = max(y)
)
}
\arguments{
\item{y}{An boolean response vector.}

\item{X}{A design matrix (full of integer covariates).}

\item{iter}{the total number of mcmc iterations.}

\item{warmup}{the number of warmup mcmc iterations.}

\item{seed}{the seed for random number generation.}

\item{rho}{the hyperparameter \eqn{\rho}.}

\item{a}{the hyperparameter \eqn{a}.}

\item{b}{the hyperparameter \eqn{b}.}

\item{a2}{the hyperparameter \eqn{a_2}.}

\item{b2}{the hyperparameter \eqn{b_2}.}
}
\value{
An object of class \code{anova_bnp_model}.
}
\description{
Fit a Bayesian nonparametric factorial ANOVA model with Bernoulli kernel
}
