% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bnp_normal.R
\name{anova_bnp_normal}
\alias{anova_bnp_normal}
\title{Fit a Bayesian nonparametric factorial ANOVA model with Gaussian kernel}
\usage{
anova_bnp_normal(
  y,
  X,
  iter = 4000L,
  warmup = 2000L,
  seed = 1L,
  n = 50L,
  rho = 1,
  a = 1,
  b = 1,
  a0 = 2,
  lambda0 = 1,
  mu0 = 0,
  b0 = 1,
  lb = mean(y) - 3 * sd(y),
  ub = mean(y) + 3 * sd(y),
  standardize_y = FALSE
)
}
\arguments{
\item{y}{a continuous response vector.}

\item{X}{a design matrix (full of integer covariates).}

\item{iter}{the total number of mcmc iterations.}

\item{warmup}{the number of warmup mcmc iterations.}

\item{seed}{the seed for random number generation.}

\item{n}{the number of points \eqn{y_0} for computing \eqn{p(y_0 | y)}.
The final grid is conformed by \code{n} equispaced points from \code{lb} to \code{ub}, see
the arguments \code{lb} and \code{ub}.}

\item{rho}{the hyperparameter \eqn{\rho}.}

\item{a}{the hyperparameter \eqn{a}.}

\item{b}{the hyperparameter \eqn{b}.}

\item{a0}{the hyperparameter \eqn{a_0}.}

\item{lambda0}{the hyperparameter \eqn{\lambda_0}.}

\item{mu0}{the hyperparameter \eqn{\mu_0}.}

\item{b0}{the hyperparameter \eqn{b_0}.}

\item{lb}{the lower bound of the prediction grid.}

\item{ub}{the upper bound of the prediction grid.}

\item{standardize_y}{Select \code{TRUE} to internally standardize the data,
fit the model, and present the results in the original scale.}
}
\value{
An object of class \code{anova_bnp_model}.
}
\description{
Fit a Bayesian nonparametric factorial ANOVA model with Gaussian kernel
}
