% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bnp_normal.R
\name{anova_bnp_normal}
\alias{anova_bnp_normal}
\title{Fit a Bayesian nonparametric factorial ANOVA model with Gaussian kernel.}
\usage{
anova_bnp_normal(
  y,
  X,
  iter = 4000L,
  warmup = 2000L,
  seed = 1L,
  n = 50L,
  zeta0 = 1,
  a0 = 1,
  b0 = 1,
  v0 = 2,
  r0 = 1,
  u0 = 0,
  s0 = 1
)
}
\arguments{
\item{y}{A continuous response vector.}

\item{X}{A design matrix (full of integer covariates).}

\item{iter}{the total number of mcmc iterations (default 4000).}

\item{warmup}{the number of warmup mcmc iterations (default 2000).}

\item{seed}{the seed for random number generation (default 1L).}

\item{n}{the number of points y0 for computing p(y0 | y) (default 50L).
The final grid is conformed by n equispaced point from
\code{min(y) - sd(y) / 2} to \code{max(y) -  sd(y) / 2}}

\item{zeta0}{the hyperparameter \eqn{\zeta_0} (default 1.0).}

\item{a0}{the hyperparameter \eqn{a_0} (default 1.0).}

\item{b0}{the hyperparameter \eqn{b_0} (default 1.0).}

\item{v0}{the hyperparameter \eqn{v_0} (default 2.0).}

\item{r0}{the hyperparameter \eqn{r_0} (default 1.0).}

\item{u0}{the hyperparameter \eqn{u_0} (default 0.0).}

\item{s0}{the hyperparameter \eqn{s_0} (default 1.0).}
}
\value{
An object of class 'anova_bnp_model'.
}
\description{
Fit a Bayesian nonparametric factorial ANOVA model with Gaussian kernel.
}
